fix https://github.com/yegor256/takes/issues/945
