+++ src/main/java/org/takes/facets/auth/social/PsFacebook.java
@@ -46,7 +46,7 @@ import org.takes.facets.auth.Identity;
 import org.takes.facets.auth.Pass;
 import org.takes.misc.Href;
 import org.takes.misc.Opt;
-import org.takes.rq.RqHref;
+import org.takes.rq.RqHrefBase;
 
 /**
  * Facebook OAuth landing/callback page.
@@ -142,7 +142,7 @@ public final class PsFacebook implements Pass {
     @Override
     public Opt<Identity> enter(final Request trequest)
         throws IOException {
-        final Href href = new RqHref.Base(trequest).href();
+        final Href href = new RqHrefBase(trequest).href();
         final Iterator<String> code = href.param(PsFacebook.CODE).iterator();
         if (!code.hasNext()) {
             throw new HttpException(
