+++ src/main/java/org/takes/facets/auth/social/PsGoogle.java
@@ -40,7 +40,7 @@ import org.takes.facets.auth.Identity;
 import org.takes.facets.auth.Pass;
 import org.takes.misc.Href;
 import org.takes.misc.Opt;
-import org.takes.rq.RqHref;
+import org.takes.rq.RqHrefBase;
 
 /**
  * Google OAuth landing/callback page.
@@ -145,7 +145,7 @@ public final class PsGoogle implements Pass {
     @Override
     public Opt<Identity> enter(final Request request)
         throws IOException {
-        final Href href = new RqHref.Base(request).href();
+        final Href href = new RqHrefBase(request).href();
         final Iterator<String> code = href.param(PsGoogle.CODE).iterator();
         if (!code.hasNext()) {
             throw new HttpException(
