+++ src/main/java/org/takes/facets/fallback/FbSlf4j.java
@@ -30,7 +30,7 @@ import org.cactoos.text.TextOf;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.takes.misc.Opt;
-import org.takes.rq.RqHref;
+import org.takes.rq.RqHrefBase;
 import org.takes.rq.RqMethod;
 
 /**
@@ -69,7 +69,7 @@ public final class FbSlf4j extends FbWrap {
         FbSlf4j.LOGGER.error(
             "{} {} failed with {}: {}",
             new RqMethod.Base(req).method(),
-            new RqHref.Base(req).href(),
+            new RqHrefBase(req).href(),
             req.code(),
             new TextOf(new BytesOf(req.throwable()))
         );
